+++ lithops/storage/backends/aws_s3/config.py
@@ -22,9 +22,6 @@ def load_config(config_data):
 
     if 'aws' in config_data:
 
-        if not {'access_key_id', 'secret_access_key'}.issubset(set(config_data['aws'])):
-            raise Exception("'access_key_id' and 'secret_access_key' are mandatory under the 'aws' section of the configuration")
-
         if 'aws_s3' not in config_data:
             config_data['aws_s3'] = {}
 
@@ -37,8 +34,3 @@ def load_config(config_data):
 
         if 'region' not in config_data['aws_s3']:
             raise Exception("'region' is mandatory under 'aws_s3' or 'aws' section of the configuration")
-
-        if 'storage_bucket' not in config_data['aws_s3']:
-            key = config_data['aws_s3']['access_key_id']
-            region = config_data['aws_s3']['region']
-            config_data['aws_s3']['storage_bucket'] = f'lithops-{region}-{key[:6].lower()}'